+++ pkg/controller/controller_instance.go
@@ -18,7 +18,6 @@ package controller
 
 import (
 	"fmt"
-	"net/http"
 	"net/url"
 
 	"github.com/golang/glog"
@@ -1168,17 +1167,6 @@ func clearServiceInstanceCurrentOperation(toUpdate *v1beta1.ServiceInstance) {
 	toUpdate.Status.LastOperation = nil
 	toUpdate.Status.InProgressProperties = nil
 	toUpdate.Status.ReconciledGeneration = toUpdate.Generation
-}
-
-// shouldStartOrphanMitigation returns whether an error with the given status
-// code indicates that orphan migitation should start.
-func shouldStartOrphanMitigation(statusCode int) bool {
-	is2XX := (statusCode >= 200 && statusCode < 300)
-	is5XX := (statusCode >= 500 && statusCode < 600)
-
-	return (is2XX && statusCode != http.StatusOK) ||
-		statusCode == http.StatusRequestTimeout ||
-		is5XX
 }
 
 // serviceInstanceHasExistingBindings returns true if there are any existing
